Starting a distributed DofManager
git-svn-id: https://geuz.org/svn/gmsh/trunk@19288 59c8cabf-10d6-4a28-b43f-3936b85956c4

+++ projects/small_fem/assembler/Dof.cpp
@@ -11,7 +11,7 @@ Dof::Dof(const Dof& other){
   this->type    = other.type;
 }
 
-Dof::Dof(size_t entity, size_t type){
+Dof::Dof(int entity, int type){
   this->entity  = entity;
   this->type    = type;
 }
